feat: separate TopicSlide from swiper
