+++ src/components/Home/TopicSwiper/TopicSwiper.tsx
@@ -1,12 +1,9 @@
 import React, { useRef, useState } from 'react';
-import { styled } from 'styled-components';
 import SwiperCore from 'swiper';
 import { Navigation } from 'swiper/modules';
-import { Swiper, SwiperSlide } from 'swiper/react';
+import { Swiper } from 'swiper/react';
 
-import { colors } from '@styles/theme';
-
-import { RightChevronIcon } from '@icons/index';
+import TopicSlide from './TopicSlide';
 
 SwiperCore.use([Navigation]);
 
@@ -51,62 +48,24 @@ const TopicSwiper = ({ children, fetchNextPage, hasNextPage }: TopicSwiperProps)
         onReachBeginning={handleReachBeginning}
         observer={true}
       >
-        {children.map((child, index) => (
-          <TopicSlide key={index} style={{ overflowY: 'auto' }}>
-            <PrevButton
-              disabled={init || prevDisabled}
-              onClick={() => {
-                swiperRef.current?.slidePrev();
-                setNextDisabled(false);
-              }}
+        {children.map((child) => {
+          return (
+            <TopicSlide
+              key={child.key}
+              init={init}
+              prevDisabled={prevDisabled}
+              swiperRef={swiperRef}
+              setNextDisabled={setNextDisabled}
+              nextDisabled={nextDisabled}
+              setPrevDisabled={setPrevDisabled}
             >
-              <RightChevronIcon style={{ transform: 'rotate(180deg)' }} stroke={colors.white_40} />
-            </PrevButton>
-            {child}
-            <NextButton
-              disabled={nextDisabled}
-              onClick={() => {
-                swiperRef.current?.slideNext();
-                setPrevDisabled(false);
-              }}
-            >
-              <RightChevronIcon stroke={colors.white_40} />
-            </NextButton>
-          </TopicSlide>
-        ))}
+              {child}
+            </TopicSlide>
+          );
+        })}
       </Swiper>
     </React.Fragment>
   );
 };
-
-const TopicSlide = styled(SwiperSlide)`
-  &::-webkit-scrollbar {
-    display: none;
-  }
-
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none; /* Firefox */
-`;
-
-const SlideButton = styled.button<{ disabled: boolean }>`
-  position: absolute;
-  top: 63px;
-  z-index: 100;
-  width: 32px;
-  height: 32px;
-  padding: 4.8px 10.4px;
-  cursor: pointer;
-  background-color: transparent;
-
-  ${(props) => props.disabled && `display: none;`}
-`;
-
-const PrevButton = styled(SlideButton)`
-  left: 20px;
-`;
-
-const NextButton = styled(SlideButton)`
-  right: 20px;
-`;
 
 export default TopicSwiper;
